+++ lib/main.dart
@@ -5,9 +5,7 @@ import 'package:book_reader/screens/login_screen.dart';
 import 'package:book_reader/size_data.dart';
 import 'package:flutter/material.dart';
 
-void main() {
-  runApp(const MyApp());
-}
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
@@ -34,7 +32,7 @@ class MyApp extends StatelessWidget {
           colorScheme: ColorScheme.fromSeed(seedColor: Colors.amber),
           useMaterial3: true,
         ),
-        home: const DetailsScreen(),
+        home: const LibraryScreen(),
       ),
     );
   }
